Persistent Objects
Project requires a command line, menu driven inventory controll project that
saves inventory as a structure and allows to add new inventory items, edit
existing items, and view items. It's a modification of an earlier project
with added concept of saving the structures to a file for persistence.

filetest.cpp is a simple file read/write prototype where I test the basic
file saving concepts one at a time in a simple environment.

inventory.cpp will be the final project. At this time it is a converted
version of the previous project and ready to handle a vector of inventory
structures. All it needs is file persistence.