LoadStr not all there
LoadStr module added. since it opens and closes the file in a single call,
LoadStr only ever returns the first line of a file.

The intent is to return each consecutive line of a file. Possibly return the
string via pointer and return the get position of a successful load which
can be passed back to the file.

Alternatively, I could make a LoadFile routine, which getlines until end of
file, storing each line in a vector or other structure.

Ultimately we want to get each line of the file and add to the members of a
structure. Then another structure once all of those lines are used.`