Beginning of delete opperations
A file delete is being added. I've learned that the different methods for
opening a file only determine where the file pointer begins, and not actualy
clears out the file, which is fine. Delete will have to be modified to get
the legnth of the file and overwrite it with NULLs or some other whitespace.

After that, I just need to decompoose my experiment from main and add my
newly created functions to my code according to the following design:

start
load the file into the vector
menu opperations
after edit complete: delete file, save the vector
after new complete: delete file, save the vector