filetest complete
enclosed are a collection of file manipulation functions that are the
beginnings of persistent objects. My deadline is looming, so I am settling
for "it works" Later, I will go back and clean these up. LoadFile needs a
lot of work to be made modular, and I would like to make these modules act
as general member functions at some point.

I will invent a persistent, self-saving class eventualy, if I have to write
it in assembley to do it.

Half of this code was written with a bucket between my legs, into which I
vomited some really acidic bile while coding. I don't reccomend the generic
can of corned beef and hash for a pre-coding breakfast.

That last part was not a joke.