Done
The stream used to convert numbers to strings and write them to the file was
accumulating each number entered. If I wrote 6 for a value, then 7 for
another value, when saved, the stream would pass 6 to the first, then 67 to
the next.

I first tried scoping the stream to the for statement, but I still had the
accumulation withing single inventory entries. Thank you Google: turns out
you can clean out a stream by calling stringstream.str("").

And with that the last bug is fixed. Sleep time. Tomorrow Im going to relax
with some easy assembly language.